LPS-147531 Create an index to reuse IObjectFieldBusinessType
