+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ModalAddObjectField.tsx
@@ -351,13 +351,6 @@ interface IProps extends React.HTMLAttributes<HTMLElement> {
 	onClose: () => void;
 }
 
-interface IObjectFieldBusinessType {
-	businessType: string;
-	dbType: string;
-	description: string;
-	label: string;
-}
-
 type TPicklist = {
 	id: string;
 	name: string;
